+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - (API) Added method `Container.isFull()` to check if a container is full.
 - (API) Added shield blocking state and cooldown accessors to `EntityPlayerBaseComponent`.
 - (API) Added `EntityPhysicsComponent.resetFallDistance()`.
+- (API) Added attacker/target overload for `ItemBaseComponent.calculateAttackDamage()`.
 - (API) Added method `BlockEntityChest.getDoubleChestContainer()` to get the container object to use when the chest is paired.
 - (API) Implemented hopper:
   - (API) Added `BlockEntityHopper` and related interfaces.